iso: Add support for selecting the type of BIOS iso
This adds an ISOBootType enum that replaces the ISOLinux bool in order
to control what type of boot iso is created:
 - UEFI only using grub2
 - syslinux on BIOS and grub2 UEFI
 - grub2 for both BIOS and UEFI

It also adds the Grub2MBR option to the xorrisofs stage.  Setting this
to the path of the grub2 hybrid boot image
(eg./usr/lib/grub/i386-pc/boot_hybrid.img) will cause the created ISO to
use grub2 to boot on BIOS systems.

This overrides the presence of the ISOHybridPath and does not use
syslinux.

In order to support this it does require that the grub2-pc-modules
package is installed.

This also adds a stage using Grub2InstStage to create the eltorito.img
needed by grub2.