blueprint: add enabled modules
Add a new field to the top level of the blueprint (like `packages`) that
allows for modular content to be selected.

Note that there was already a `modules` field but it *doesn't* refer to
modules, it refers to normal packages. Modules are implemented in the
new `enabled_modules` field to make sure we keep backwards
compatibility.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>