modules: pipe through workload into os
Sets enabled modules coming from customizations on the workload object
which is then merged with the rest of the package sets in the os
pipeline, which is the only pipeline where modularity matters right now.

Note that we might want to support modularity outside of workloads later
on, but not right now.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>