manifest,osbuild: handle all fs-related kargs in GenImageKernelOptions()
Move the root UUID discovery from the OS pipeline into the
GenImageKernelOptions() function.  Previously this was done in the
prependKernelCmdlineStage().  That function now simply generates the
stage and adds it to the top of the pipeline and doesn't do any
discovery on its own.

The following commits will add more filesystem-related kernel
command-line argument handling to the osbuild function.