osbuild: Add iso9660 support to grub2_inst_stage
The grub2 BIOS iso needs an eltorito.img built by grub2-mkimage. This is
created when location isn't passed to org.osbuild.grub2.inst and the
filesystem is set to iso9660.

In order to support this some of the fields become optional, and iso9660
is added to the list of allowed filesystems.