packagesets: introduce `IMAGE_BUILDER_EXPERIMENTAL=yamldir=...`
This commit adds a new `IMAGE_BUILDER_EXPERIMENTAL=yamldir=...`
experimental flag to force setting the datadir for the yaml
directory to override e.g. packagesets.

Ideally we would pass this all the way down the stack via a
search path but this will require a bunch of more refactoring.
So to give people the ability to tweak packagesets today, start
here.