distro: tweak the packagesets.Loader() interface
The current `packagesets.Loader()` takes a single imagetype
argument to determine what packagesets to load from YAML.

This generally works great but its a challenge for image types
like the anaconda installer where we need a different package
set for the iso and the os. In fedora we workaround this by
re-using the minimal-raw package set for the anaconda installers
"os" package set. However in Centos/RHEL this is less straightforward
so having the ability to "override" the pkgset name is required.