many: make bootstrap container usage a manifest option
This commit adds the option to use a bootstrap container when
generating the manifest. This allows a seamless experience when
using e.g. image-builder:
```console
$ image-builder build --arch=riscv64 minimal-raw --distro fedora-42
```

The approach is to make the distro bootstrap container ref part of
the manifest struct just like we set "Distro" there. This is not
ideal, it would be much nicer if manifest would hold the full
`distro.Distro` and this way we could simply run something like
`distro.BootstrapContainerRef()` when needed. However because
of circular inputs that is not possible.

The other option is to make it part of every `imageFunc()` call,
this was suggested by Achilleas (thanks!) this requires some
refactor. Or we could make it part of `InstantiateManifest()`
and pass a new generalized `RepoAndContainersConfig` there that
would contain both the []rpmmd.RepoConfig and the BootstrapRef.

Note that I think eventually the bootstrap container ref will be
part of the "distro.yaml" and the currently (slightly ugly) helper
`bootstrapContainerFor()` will go away.