many: make bootstrap container usage a manifest option
This commit adds the option to use a bootstrap container when
generating the manifest. This allows a seamless experience when
using e.g. image-builder:
```console
$ image-builder build --arch=riscv64 minimal-raw --distro fedora-42
```

The approach is to make the distro bootstrap container ref part of
the manifest struct just like we set "Distro" there. This is not
ideal, it would be much nicer if manifest would hold the full
`distro.Distro` and this way we could simply run something like
`distro.BootstrapContainerRef()` when needed. However because
of circular inputs that is not possible.

The other option would be to change the signature of
image.ImageKind.InstantiateManifest, e.g.
```go
- InstantiateManifest(m *manifest.Manifest, repos []rpmmd.RepoConfig, runner runner.Runner, rng *rand.Rand) (*artifact.Artifact, error)
+ InstantiateManifest(m *manifest.Manifest, repos newpkg.RepoAndBootrapConfig, runner runner.Runner, rng *rand.Rand) (*artifact.Artifact, error)
```
here it would be either a new `rpmmd.RepoAndBootstrapConfig` with a struct
that holds []rpmmd.RepoConfig and BootstrapRef - but that feels a bit
out of place in rpmmd (because of the container ref that is unreleated to
rpm repos). The RepoAndBootstrapConfig cannot be part of manifest,image
or distro because of circular imports.

Alternatively:
```go
- InstantiateManifest(m *manifest.Manifest, repos []rpmmd.RepoConfig, runner runner.Runner, rng *rand.Rand) (*artifact.Artifact, error)
+ InstantiateManifest(m *manifest.Manifest, repos []rpmmd.RepoConfig, bootstrapRef string, runner runner.Runner, rng *rand.Rand) (*artifact.Artifact, error)
```
that just adds a boostrapRef arg. That seems ok but requires a
lot of (mechanical) changes.

Note that I think eventually the bootstrap container ref will be
part of the "distro.yaml" and the currently (slightly ugly) helper
`bootstrapContainerFor()` will go away.