packagesets: add basic imagetype yaml to merge pkgSets
This commit adds a basic imagetype yaml to support merging
of package sets  This is a useful feature so that we can e.g.
avoid duplicating the `cloud_base` or `anaconda` package sets.

It works by adding a new toplevel "image_types" in the
package_sets.yaml.

It allows us to write:
```yaml
image_types:
 qcow2:
  package_sets:
   - cloud_base_pkgset
   - qcow2_pkgset

package_sets:
  cloud_base_pkgset:
    include: [foo]
    exclude: [bar]
  qcow2_pkgset:
    include: ["qemu-guest-agent"]
```
The downside of this approach is that it is nice to read but not
symetrical, i.e. one cannot just include the pkgset verbatim in
an `image_type`, it always needs to be defined inside the
"package_sets" key. It also uses the same "package_sets" name
for two different things which is also slightly not nice.