test: add new manifest reference tests
This commit adds new manifest reference tests to capture when we
core aspects of the generated manifests. This is inspired by the
work we did in `otk` [0].

The test is very fast, there is no depsolving or other network
activity, but because "fake" packages get generated we will
notice any changes in the packagelists.

This is very useful when working on the yamlification of the
package lists (and later the other parts) because this way
we will spot any differences right away.

[0] https://github.com/osbuild/otk/blob/main/test/test_against_images_refs.py