distro: make yamlloader deal with different dir layouts
This commit makes the yaml loader deal with different layout
for the yaml directories and also deals with the fact that
go:embed does not support symlinks.

This is (hopefully) a temporary measure, I expect that eventually
there will be a `distros.yaml` or similar file that describes
the various distros supported and some details (like their data
dir).