manifest: rename OS pipeline ExtraBasePackages to BasePackages
In the OS pipeline, the word Extra in ExtraBasePackages is a bit
misleading.  The property is used to add the static image type package
set to the pipeline.  They are 'extra' in the sense that they are
included in addition to the auto-selected packages required by the
pipeline and its customizations, but IMO the wording makes it seem that
they are a package set that's included from some other process separate
from the image type and pipeline.