manifest: remove OSTreeRef check for partition table packages
OSTree-based images don't have a partition table so the check is
redundant and potentially confusing.
The comment about LVM is also very outdated and originates from the time
when lvm packages were added based on a plain 'HasLVM' boolean on the
pipeline.

The code is pretty self-explanatory now.

This change has no effect on generated manifests.