distro: rename "packagesets.Load" to "defs.PackageSet"
This commit renames the existing `packagesets` module to `distro/defs`
to reflect that it will be about more than just the packagesets.

There are various other names we could use, I have no super strong
opinion but `distro/defs` is nice and short. Some more options:
- loader.{Load,}PackageSet
- yamlload.PackageSet
- distrodefs.PackageSet (distro/distrodefs feels a bit redundant)
- defload{,er}.PackageSet()

This prepares for the upcoming:
- defs.PartitionTable
- defs.ImageConfig
- defs.Platform
helpers.