manifest: add bootstrap support to the buildroots
This commit adds a new `manifest.NewBootstrap()` helper and an
optional `BuildOptions.BootstrapBuildroot` that can be used
to create a buildroot that uses a bootstrap buildroot to get
created.

The use-case is cross-arch/cross-distro builds where we use a
bootstrap container to create the actual buildroot. The requirements
on the bootstrap container are very lightweight and any standard
container with rpm/python3/mount should do it. With those osbuild
can install the real buildroot that will contain mkfs all the
user tools it needs.