distro: move `osbuild.WSLConfStageOptions` out of ImageConfig
In PR#1337 the desire was expressed to move all direct `osbuild`
reference out of `distro.ImageConfig` to make it a more high
level abstraction. This commit is the first step to show how
this could be done with the simple `osbuild.WSLConfStageOptions`.

However doing this with the more complex ones will be quite a
bit of busywork and reviewing. I am not sure we should do this
before we do the yamlification as I'm worried converting the
other ~26 structs that needs similar converstion will take some
time.

The other concern with this is that we currently have json schema
information for all osbuild options. So validating the yaml against
the json-schema of osbuild seems feasible. If add this conversion
stage we will need to also duplicate the json schema.