osbuild: do not add FIPS files when generating FIPS stages
The GenFIPSStages() function would also add file node stages for files
related to FIPS.  This commit removes the file node stages from the
generated list.

In the OS pipeline generator this means we now have to add the stages
explicitly to the pipeline.  This is because the FIPS stages are
generated *after* the handling of the Files list form the
OSCustomizations, which we generally use to add files to the end of the
pipeline.

In the OSTreeDeployment pipeline, this change removes duplicate stages.
The FIPS file stages were being added twice because the files were being
added both to the pipeline stages directly and to the OSCustomizations
Files list *before* the handling of the Files list.