manifest: addInlineDataAndStages() helper method for the OS pipeline
The new addInlineDataAndStages() method of the OS pipeline generator
appends file creation, chmod, and chown stages to the pipeline and also
adds the inline data to the inlineData property of the OS.  The
inlineData list is used to return inline data that was collected during
manifest generation.

Previously, we used the Files list in two ways:
- To define files to be created externally, during the initialisation of
  OSCustomizations (e.g. from a blueprint).
- To collect files that needed to be created under certain conditions or
  for certain stages while generating the pipeline.

This lead to inconsistencies.  The handling of the Files list was meant
to always be near the end of the pipeline, before the selinux stage.
However, files were being added to the list after the Files list was
handled, which meant that appending to the Files was done to generate
the inline data, but not to generate stages.

This change also makes the usage of OSCustomizations more consistent.
During pipeline generation, it's better to treat OSCustomizations as
read-only properties, initialised before serialization starts.
We might consider enforcing this in code.