manifest: addInlineDataAndStages() for OSTreeDeployment pipeline
This is the same change as the previous commit adapted for the
OSTreeDeployment pipeline.  The main difference is that file stages need
to mount the ostree deployment.

Manifest changes:
This commit modifies the generated manifests that have an
OSTreeDeployment pipeline and enable FIPS.  The file creation stages
have moved up in the pipeline now, but functionally it should make no
difference.