fedora: kernel opts to be `rw` for minimal
Currently on Fedora 43 we only write mount units, without an fstab. This
causes the system to boot with read-only `/`. I haven't quite figured
out the why (systemd perhaps looks at the kernel options, perhaps its
the default in mount units and needs to be explicitly enabled).

Regardless, it seems that we want the kernel opts to be `rw` and to not
rely on things to remount `/` later on.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>