defs: add `partition_table_override` to `distro.yaml`
With the most recent changes for the `iot-raw` imagetype
in fedora we need a way to conditionally override parts
of the partition table.

This commit adds this support via a simple:
```yaml
    partition_table_override:
      condition:
        version_greater_or_equal:
          "42":
            - select_partition: 2
              fstab_options: "defaults,ro"
```
It currently supports `size` (we will need this for RHEL) and
`fstab_options`. It is sadly not generic (we can look into
this using reflection) and a bit ad-hoc, i.e. the selector
is currently very primitive and will fail for nested payloads
(like overrides for a FS inside LUKS). We have no need for
the nesting yet though and if it comes up we could invent a
notion like (`select_partiton: 2:1` or similar /if/ we need
it).