Test/manifests: check podman network_backend consistency
When embedding containers into images, certain versions of podman
may consider them as a sign of system upgrade or migration and may
fall back to using `cni` network_backend for backward compatibility
even though it should use its default. Since we embed containers as
root, this manifests as different network_backend being used for
rootfull and rootless podman.

Add a check to verify that the network_backend is the same for rootfull
and rootless podman when embedding container into the image.

More information in:
https://docs.podman.io/en/v4.2/markdown/podman-network.1.html

Signed-off-by: Tomáš Hozza <thozza@redhat.com>