defs: add new `partition_mount_point` select to overrides
The rhel8 partition table is slightly more complicated than rhel7
and rhel10. It requires changing the size of the "/boot" partition
depending on the version but only aarch64 systems have a /boot
partition so the existing selection based on partition index to
override does not work. Hence a new `partition_mount_point` selector
is added that can override:
```yaml
    partition_table_override:
      condition:
       version_less_than:
          "8.9":
            - partition_mount_point: "/boot"
              partition_mode: ignore-missing
              size: 222_222_222
```
It also adds the ability to control how the selection works, i.e
if not finding the selected partition is an error (the default)
or should be ignored (which we need for rhel8).