distro: add `ImageType.BasePartitionTable()`
This commit exposes the base partition table for `ImageType`s.
The rational is that we want to expose those via:
```console
$ image-bulder describe-image qcow2
```
so that the yaml of the partition table is visible. This helps
developing our image definitions wihtout having to look at the
code and as added benefit it gives a "flattened" version of the
partition YAML.

As part of the work in fedora the `basePartitionTablesFunc` was
removed as it is now fully based on YAML. Once all partition
tables in RHEL are also in YAML [0] we can do a similar
simplification there.

[0] https://github.com/osbuild/images/pull/1372