distro: special case Alma Linux Kitten
When calling `GetHostDistroName` then we need to special case Alma Linux
Kitten to avoid misidentifying it as `almalinux-10`. When the major
component is 10 and there is no minor component we need to switch to
using `kitten` as the identifier.

This is similar to CentOS/RHEL except that the switching isn't done on
the actual `"ID"` but rather the combination of it with the
`VERSION_ID`.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>