anaconda_tar_installer: Make kickstart and liveimg.tar.gz optional
This makes the Anaconda tar installer an optional extension to the basic
installer netinst iso. If img.Kickstart is not set it does not include
any kickstart and acts like a basic netinst boot.iso

Callers of NewAnacondaTarInstaller must set img.Kickstart if they want
to embed a kickstart. All current callers already do this.

This adjusts the tests to account for this new behavior and updates the
manifest-checksums.txt