manifestgen: extra io.Writer for depsolver
Let's pass along a DepsolveOutput field all the way to the depsolver
so we can use it for the stderr of the `dnf-json` command. This then
allows us to capture its output (mostly: package in group but excluded
warnings) in, for example, `image-builder-cli`.

The depsolver already had provisions for passing in a writer for
`os.Stderr`.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>