distro: unify OsCustomizations()
This commit removes the two (almost) identical copies of
`OsCustomizations()` in {fedora,rhel} in favor of a single
implementation that takes the new ImageTypeConfig in additon
to the existing parameters.

This (hopefully) goes into the direction that Achilleas suggested
where we have a new single SuperConfig that is produced via
a smart merge of `ImageConfig`, `ImageTypeConfig`, `DistroConfig`,
Blueprints etc. But in order to make progress on this journey
we need to remove some duplication first :)

Along the way a subtle difference in the hostname handling was
unified and RHEL now also understands about `MachineIdUninitialized`
(which was previously only available for fedora probably because
of drift).

There is a tradeoff here of course, its slightly^W ugly that now
distro.OsCustomizations needs to understand about the details
of *all* distros/imageTypes we support. We could explore extracing
a common helper and then let the distro deal with specialized stuff
like subscriptions. But moving into this (other) extreme seems
a slightly better trade-off than the current duplicated ~300 lines
that we need to keep drift-free.