defs: make condition(s) version sorted
When we go over conditions that require versions we need to make
sure we sort them first so that if there are multiple version
conditions the way they are applied is predictable. As they
are just map keys in YAML the order is not preditable otherwise.
So with
```
condition
 version_greater_or_equal:
   "1":
     - foo
   "2":
     - bar
```
we need stable ordering to ensure that when e.g. a partition
table is overriden they are executed in order so that the last
change can potentially override the previous one.

Note that this version requires go1.23 which we currently cannot
use. So there is an extra commit to provide a go1.22 version that
we can just revert once we can use go1.23.