dnfjson: provide details if subscription cannot be found
This commit tweaks the way error from subscription are handled
in dnfjson. Current any error details are ignored when trying
to load a subscription. But this is problematic as it makes
diagnosing issues like:
https://github.com/osbuild/image-builder-collection/issues/3
more difficult.

So this commit includes the error details of trying to load
the subscriptions in the error message of dnfjson.