osbuild: update chrony stage with new refclock options
Update the chrony stage to support the new refclock options.

The validation reuses the invalid path regular expression from the chown
stage.  The variable has been renamed to reflect its general use and
moved to the main file of the package (osbuild.go).

Each refclock driver has its own constructor that takes as argument only
the mandatory option for the driver and sets the proper name on the
struct.