installer: Change installer dracut use to match Lorax
dracut should be in charge of the modules installed in the inird. This
ensures that changes needed to enable new hardware support don't need to
chase down every project that uses it. We do this by using --no-hostonly
mode which creates an initrd that will boot on any hardware we support.
Note that we don't need to pass this to it, the osbuild stage always
includes it when calling dracut.

Image creation tools like lorax and osbuild should only added modules
needed to support the boot environment (eg. anaconda or dmsquash-live).

This commit refactors things to accomplish this, removing the large
hard-coded list of dracut modules and building a dracut stage that
matches how it is used when building the boot.iso