installer: Change installer dracut use to match Lorax
dracut should be in charge of the modules installed. This ensures that
changes needed to enable new hardware support or dracut changes don't
need to chase down every project that uses it. We do this by using
--no-hostonly mode while creates an initrd that will boot on any
hardware we support.

Image creation tools like lorax and osbuild should only added modules
needed to support the boot environment (eg. anaconda or dmsquash-live).

This commit refactors things to accomplish this, removing the large
hard-coded list of dracut modules and building a dracut stage that
matches how it is used when building the boot.iso