tools/gen-manifest-checksums: sort by image name
Instead of sorting the file by checksum, sort by the image name
(`--key 2`, the second field of each line).

This makes it easier to read diffs because the filenames don't move
around when their checksum changes.  It makes it much easier to
distinguish cases where image configurations are added or removed, as
opposed to images changing.

The script sets LC_ALL=C to make sure the sorting is stable.  From the
sort manpage:

  *** WARNING *** The locale specified by the environment affects sort
  order.  Set LC_ALL=C  to  get  the  traditional sort order that uses
  native byte values.