distro: rename ImageConfig SystemdUnit to SystemdDropin
The org.osbuild.systemd.unit stage creates drop-in systemd files for
modifying existing systemd units.  It predates the
org.osbuild.systemd.unit.create stage, which is meant to create whole
systemd unit files.
We're about to add whole systemd units to the ImageConfig, so to avoid
confusion and to make the purpose of each option clearer, rename the
SystemdUnit ImageConfig property to SystemdDropin, so that the new
option can be called SystemdUnit.