distro/rhel9/azure: change console kernel args for aarch64
This aligns the aarch64 Azure images with the current marketplace images
for RHEL 9.5 and older.  The change drops console=tty1, console=ttyS0,
earlyprintk=ttyS0, and rootdelay=300 from the ARM (aarch64) kernel args
and adds console=ttyAMA0.

Unlike other changes we've been making to the Azure images, this is not
version-gated to 9.6+.  Our previous kernel args cause issues with
console output for aarch64 images, so it's better to change them across
the whole range.

We usually don't have image type initialisers that take an architecture
argument, so this approach of passing the architecture directly to the
initialiser is a bit quick and simple.  A more general solution, one
which might also support arch-specific image configs would be nicer.
Something similar to the partition table selection, which is a function
property of the image type, might be needed.

Partially addresses #1455.