fedora: server-wsl stays as wsl
Due to some issues in testing the manifests it is problematic to have
this image type be named `server-wsl`. Since this keeps blocking the
entire PR I'd prefer to rename it back to `wsl`, and then reintroduce
the change/revert in a separate PR where the CI issue(s) can also be
tackled.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>