base-host-check: hostname is a warning
Let's make the hostname check emit a yellow circle as a warning, it's
currently kinda expected in the image types that we boot test to have
this fail due to cloud-init setting the host name on boot.

When we start boot testing more image types this can be conditionalized
on that.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>