distro: make `DNFConfig` more declarative
This commit removes the `DNFConfigStageOptions` from
`ImageConfig` in favor of a simpler `DNFConfig` struct.

The rational is that `DNFConfigStageOptions` current gets
the `osVersion` passed in directly. When we move to a
YAML based ImageConfig this will be problematic as we
do not have access to variables like this. By making this
a declarative option we sidestep the problem.