distro: add DNFSetReleaseVerVar to ImageConfig
This commit adds a new `DNFSetReleaseVerVar` to
`ImageConfig` and a new `DNFConfigOptions() helper.

The rational is that `DNFConfigStageOptions` current gets
the `osVersion` passed in directly. When we move to a
YAML based ImageConfig this will be problematic as we
do not have access to variables like this. By making this
new declarative option we sidestep the problem.

This option is used in the rhel sap images.

Its slightly unfortunate that we need to change two
places (images.go in rhel/fedora) to use the new
helper (pr#1387 would fix this) but with more YAML
work eventually this should converge.

Thanks to thozza for suggesting this simplified approach.