distro: move `kernelOptions` into `ImageConfig`
This commit moves the `imageType.kernelOptions` option into
`distro.ImageConfig`. The rational is that when we move to
YAML we will need to be able to override the kernel options
based on various conditions.

Here is an example from mkMinimalRawImgType():
```go
if common.VersionGreaterThanOrEqual(d.osVersion, "43") {
	it.kernelOptions = []string{"rw"}
	}
```
The conditions will be implemented on the imageConfig but
having conditions when defining image types would be a
bit stange. The easiest fix is to move kernelOptions into
ImageConfig. Fwiw, I think conceptually it also makes
sense there so this is probably a worthwhile chage on
its own but it will make the YAML work in
https://github.com/osbuild/images/pull/1462
easier.