fsnode: remove Getters, export field to make JSON friendly
When we move `ImageConfig` into YAML we will need support for
`fsnode.File` as it is used as part of the ImageConfig.

This can be done in multiple ways, one is to add a custom
unmarshaller. This is done in PR#1479 but because of the
way the code was written the custom unmarshaling is rather
ugly.

This commit is an alternative version. As a starting point
the struct fields are simply exported and the getters of
the same name removed. Then we can add a custom unmarshal
that just validates.