policies: allow `/boot/efi` for disk customizations
With the new customization support for `/boot/efi` we will
need to allow `/boot/efi` as a valid custom mountpoint for
disk customizations.

The disk customizations will already double check that
`/boot/efi` is the right filesystem type (vfat) so allowing it
for this should also be fine.

However we cannot allow this for filesystem customizations
yet. Manual testing shows that the code does not seem DTRT
so we disallow it for filesytem customizations still until
this gets fixed.