os: Fix race condition on OSCustomizations.Grub2Config
This is a result of the overuse of pointers to data structures. In this
case the Grub2Config struct is reused and during parallel testing it
revealed that the Default field was being changed by multiple
goroutines.

Work around this by making a copy of the struct before changing the
Default string and using it in the pipeline.