distro/rhel9/azure: drop net.ifnames=0 kernel arg
For compatibility with earlier releases, we decided not to add
net.ifnames=0 to all RHEL 9.6 Azure images.
- In 1p (Azure first party marketplace) images, the kernel argument was
  present in older versions.
- In 3p (third party) images, it was not.

In order to maintain minor-version upgrade compatibility, we will
maintain this inconsistency for the rest of the RHEL 9 images.  We will
rely on pungi configs to add the argument to the 1p images going
forward.

This reverts commit 9b71e183c1b2aee834279d03cd232e54ab7b350d.

See RHEL-89440