fsnode: add custom unmarshalers for File/Dir
This commit adds custom unmarshalers for fsnode.{File,Directory}
that also do validation. For convenience when unmarshaling a
`fsnode.File` we support an extra `text: foo` field so that
in an ImageConfig YAML we can write:
```yaml
minimal_raw: &minimal_raw:
   image_config:
     files:
       - path: "/root/anaconda-ks.cfg"
         user: "root"
         group: "root"
         text: |
           # Run initial-setup on first boot
           # Created by osbuild
           firstboot --reconfig
```
The user can also use `data:` to put a base64 encoded string
in for binary files.