distro/rhel9: make /boot 1 GiB everywhere
Users are still running into issues with too small /boot on older EL9
minor versions. Let's just bump the size retroactively, and make all
EL9 images have 1 GiB /boot. I know that this is a somewhat surprising
change, but the too small /boot is an extremely hard thing to fix
in existing images, so I think it's justified for.