manifest/os: immediately create directories
Instead of appending the directories that are necessary to the OS
customizations we create them immediately.

Punting the directories into the customizations means that the `mkdir`
stage for these is inserted later in the process. Since the copy stages
for the files are created immediately and the directories are deferred
it's possible for image types that don't have the directory pre-created
to fail to build as the copy fails.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>