distro: make `defs.PackageSets()` return a map
This commit changes the defs.PackageSet loader to return
the full map of all static package sets defined for the
pipelines. This is needed when we want to move all of
the image types into YAML. We currently define in code
what image types load from os, container or installer
pipelines but that does not translate to pure YAML.

This maps "pkgsKey" to their package sets. The
map key here is a string that can either be:
- "os": packages for the os
- "installer": packages for the installer
- "container": extra package into an iot container

- "blueprint": unused AFAICT
- "build": unused AFAICT
Note that this does not directly maps to pipeline names
but we should look into making it so.

Cleanup of https://github.com/osbuild/images/pull/1291