loader: constrain the conditions for InstallerConfig to one
This commit constraints the conditions for the InstallerConfig
to one for now. If multiple ones are used an error is raised.

We can relax this later but we need to think what it means,
right now if multiple conditions are used and multiple would
be meet the order of the implementation would decide which
one is taken.

Take this YAML:
```yaml
image_types:
  test_type:
    installer_config:
      additional_dracut_modules:
        - base-dracut-mod1
      additional_drivers:
        - base-drv1
      condition:
        version_less_than:
          "2":
            additional_dracut_modules:
              - override-dracut-mod1
        distro_name:
          "test-distro":
            additional_dracut_modules:
              - other-override-dracut-mod1
```
here it is not clear what to do if both the "version" and
"distro_name" condition are meet at the same time. Should we
merge? Was this a mistake by the user? Should one take precedence?
To avoid having to answer this now add this error.

Note that we have a similar quesiton in ImageConfig and there
the `distro_name` takes precdence currently because it's
expedient for rhel/centos (as the centos version is always 9.0
so version_less_than is problematic here). A way out might be
that version_less_than would need a distro prefix?