distros: provide a way to define the distros in pure YAML
This commit adds the ability to define distros in pure
YAML. The goal is to get rid of the `distros/fedora` directory
and to be able to fully define arbitrary distros in YAML.

This is modelled after the distrofactor and the reporegistry
handling. Here the distro provides a factory method and the
reporegistry has the known distros. In the future I would
love to see the `distros.yaml` to be the canonical reference
for all supported distros and their repos. However the current
code assumes that dropping in a `$distro-$version.json` into
/etc/osbuild-composer/repositories will be enough to create
a new distro (and limit the existing distros) so we go
this route for now for better compatibility with the exiting
code.

Note that for a new development version of fedora we need
to change `distros.yaml` now and mark a new version as the
preview release. But that is the same as now where the
pkg/distros/fedora/version.go needs to be adjusted.

For a preview of the handling of repos in distros.yaml see:
https://github.com/osbuild/images/compare/main...mvo5:yaml-distroconfig?expand=1

Moving there will also mean thinking about the "old" way
to define/limit what distros are available, i.e. the composer
rpm is using /etc/osbuild-composer/repositories to limit what
distros are buildable.