distros: provide a way to define the distros in pure YAML
This commit moves the `fedora` distro definitions into pure
YAML. The goal is to get rid of the `distros/fedora` directory
and to be able to fully define arbitrary distros in YAML.

Note that this introduces a small maintenance regression
(temporarely): when adding a new fedora version it used
to be enough to just drop a "fedora-$ver.json" file into
the `data/repositories` dir and the factory will take care
of providing the distro. With this new YAML 2 extra lines
in `defs/distros.yaml` are needed. This will eventually
become (almost) as simple as before when we merge the
default repositories into `distros.yaml` too (but we
are not there yet, for a preview see:
https://github.com/osbuild/images/compare/main...mvo5:yaml-distroconfig?expand=1
Moving there will also mean thinking about the "old" way
to define/limit what distros are available, i.e. the composer
rpm is using /etc/osbuild-composer/repositories to limit what
distros are buildable.