distro: rename "fedora" -> "generic", make loader generic
The "distro/fedora" is finally no longer fedora specific,
it takes all of the data from the distros.yaml and the
<distro_name>/distro.yaml now. Reflect this by renaming
"fedora" -> "generic" and update the defs loader to be
able to load arbitrary distro names and using the
`defs_path` property.

Note that this is most likely not the final layout of
the code, having pkg/disro/defs and pkg/distro/generic
may not be necessary and it might be simpler to just
unify them into a single pkg/distro/$name or even just
pkg/distro. OTOH having the distro "images" functions
in their own pacakge might be nice, so we need to
explore this a bit more.