fedora: implement setting of the RootfsType via YAML
This commit moves the setting of the rootfs type into the
ImageConfig and by implication into pure YAML.

The relevant image defintions are now also updated. This
also means that one can define `iso_rootfs_type: "erofs"`
trivially now per image type.

Note that this also shows an inconsistency with the old
code, some of the iso images have been moved to the new
SquashfsRootfs but other have not. We should probably
look into this too but for now this PR opts to not change
the behavior just moves the place how the RootfsType for
the iso is set.