disk: rename `MkfsOption.ToString` to `MkfsOption.String`
This commit renames `MkfsOption.ToString` to `MkfsOption.String`
and expands the test to cover that too.

In go its much more common to just have `String()` methods
to convert a type to a string. The `ToString` is very rare,
another advantage of the rename is that `%s` will automatically
work if there is a `String()`.