gen-manifests: add support for new `skip-distros` key
This commit extends the config-map.json with a new
`skip-distros` key. This can be used to temporarily skip
some distros from testing. Here is an example usage:
```json
  "./configs/iot-ostree-pull-device.json": {
    "skip-distros": [
      {
	"name": "fedora-42",
	"reason": "Issue in coreos-installer dracut module: https://github.com/coreos/coreos-installer-dracut/issues/40",
	"date": "2025-05-21T00:00:00Z"
      }
    ],
    "image-types": [
      "iot-simplified-installer"
    ]
  },
```
If a skip-distro is used for more than 90 days it will panic,
then we either need to refresh the "date" or fix the issue.