fedora: add `anaconda.ModuleUsers` to ImageInstallerImage
This commit brings the fedora:imageInstallerImage() and the
rhel:ImageInstallerImage() closer together. in RHEL we always
add the `anaconda.ModuleUsers` but in fedora we don't do this.

By having them behave the same we will have an easier time
unifying them when we move everything into YAML.

Note that I lack the domain knowledge to judge if this is the
best solution. We could also do what the iot installer does
instead:
```go
if len(img.Kickstart.Users)+len(img.Kickstart.Groups) > 0 {
	// only enable the users module if needed
	img.AdditionalAnacondaModules = append(img.AdditionalAnacondaModules, anaconda.ModuleUsers)
}
```
Suggestions welcome (but both helpers really should behave the
same).