common: extra `common.SplitDistroNameVer()` helper
This commit moves the helper to split a `distro-ver` string
into its components into `common`. We are not using the
distroidparser as it is too specialized and with pkg/distro/rhel
going away we need to rethink the whole concept. It might be
interesting to provide a way in the `DistroYAML` to specify
a regexp with named options like: `(?<name>centos)-(?P<major>)`
to provide similar functionality. But for now keep it simple,
all our distros follow the same pattern when it comes to
distro name/ver.

Note that this helper could also be called common.RSplit2() as
its really just doing this. But by calling it `SplitDistroNameVer`
we can add more semantics later, e.g. we could make sure that
after the last `-` there is a semver or similar.