generic: merge the distro tests and rename to `fedora_test.go`
This commit merges the `distro_interal.go` and `distro_test.go`
into a single file and renames it to `fedora_test.go` as this
is really testing fedora specific bits.

This place for these tests is now a bit strange, i.e. why do
we have fedora specific tests in `generic` and I think we need
to think a bit about this. I can see a `pkg/distro/internal/tests`
dir or something. But for now this should be fine and we can
fold the rhel specific tests here too for now.