distro: move `boostrap_containers` handling into YAML
This commit moves the loading of the bootstrap_containers into
pure yaml and also adds a test as this is currently not tested
in images (only in image-builder-cli).

Having a reverse dependency test that would build and run the
tests for ibcli for each images release would be great. Or
we just merge ibcli and images :)