pkg/datasizes: add helper for parsing size string in raw JSON
Add a helper function which enables one to parse a size field string
value with unit and convert it to a number representing bytes. The
function operates on mappings represented as raw JSON data ([]byte) and
returns the processed copy of the raw data.

This will be useful for adding support to use size strings with units in
our YAML definitions, instead of literal bytes value.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>